+++ fixtures/textures/splatting.jpg
